Rename: page layout `SubscribedColonies` -> `Complex`
