+++ src/modules/pages/components/RouteLayouts/Complex.ts
@@ -1,0 +1,3 @@
+export const main: string;
+export const coloniesList: string;
+export const content: string;
